Move audio types to storage instead
