+++ src/voxkit/analyzers/audio_format_profile.py
@@ -18,8 +18,9 @@ from collections import Counter
 from pathlib import Path
 from typing import Any, Dict, List
 
+from voxkit.storage.constants import SUPERSET_AUDIO_EXTENSIONS
+
 from .base import DatasetAnalyzer
-from .constants import SUPERSET_AUDIO_EXTENSIONS
 
 logger = logging.getLogger(__name__)
