--- src/voxkit/analyzers/constants.py
@@ -1,2 +1,0 @@
-# Full set of audio extensions allowed for datasets
-SUPERSET_AUDIO_EXTENSIONS = {".wav", ".flac", ".mp3", ".ogg", ".m4a"}
